some code cleanup

+++ src/Extension/ModuleExplorerPanel/GraphTopologyItem.ts
@@ -6,6 +6,7 @@ import {
 import { GraphTopologyData } from "../Data/GraphTolologyData";
 import { IotHubData } from "../Data/IotHubData";
 import { Constants } from "../Util/Constants";
+import { ExtensionUtils } from "../Util/ExtensionUtils";
 import Localizer from "../Util/Localizer";
 import { Logger } from "../Util/Logger";
 import { TreeUtils } from "../Util/TreeUtils";
@@ -20,11 +21,11 @@ export class GraphTopologyItem extends vscode.TreeItem {
         public iotHubData: IotHubData,
         public readonly deviceId: string,
         public readonly moduleId: string,
-        private readonly _graphTopology: MediaGraphTopology,
+        private readonly _graphTopology?: MediaGraphTopology,
         private readonly _graphInstances?: MediaGraphInstance[],
         private _nameCheckCallback?: (name: string) => boolean
     ) {
-        super(_graphTopology.name, vscode.TreeItemCollapsibleState.Expanded);
+        super(_graphTopology?.name ?? "", vscode.TreeItemCollapsibleState.Expanded);
         this.iconPath = TreeUtils.getIconPath(`graph`);
         this.contextValue = "graphItemContext";
         this._logger = Logger.getOrCreateOutputChannel();
@@ -49,8 +50,11 @@ export class GraphTopologyItem extends vscode.TreeItem {
         return this._instanceList;
     }
 
-    public editGraphCommand(context: vscode.ExtensionContext) {
-        const createGraphPanel = GraphEditorPanel.createOrShow(context.extensionPath, Localizer.localize("editGraphPageTile"));
+    public setGraphCommand(context: vscode.ExtensionContext) {
+        const createGraphPanel = GraphEditorPanel.createOrShow(
+            context.extensionPath,
+            Localizer.localize(this._graphTopology ? "editGraphPageTile" : "createNewGraphPageTile")
+        );
         if (createGraphPanel) {
             createGraphPanel.waitForPostMessage({
                 name: Constants.PostMessageNames.closeWindow,
@@ -59,7 +63,7 @@ export class GraphTopologyItem extends vscode.TreeItem {
                 }
             });
 
-            createGraphPanel.setupInitialMessage({ pageType: Constants.PageTypes.graphPage, graphData: this._graphTopology });
+            createGraphPanel.setupInitialMessage({ pageType: Constants.PageTypes.graphPage, graphData: this._graphTopology, editMode: !!this._graphTopology });
 
             createGraphPanel.setupNameCheckMessage((name) => {
                 return this._nameCheckCallback == null || this._nameCheckCallback(name);
@@ -74,7 +78,9 @@ export class GraphTopologyItem extends vscode.TreeItem {
                             createGraphPanel.dispose();
                         },
                         (error) => {
-                            this._logger.logError(`Failed to save the graph "${topology.name}"`, error); // TODO. localize
+                            const errorList = GraphEditorPanel.parseDirectMethodError(error, topology);
+                            createGraphPanel.postMessage({ name: Constants.PostMessageNames.failedOperationReason, data: errorList });
+                            this._logger.logError(`${Localizer.localize("saveGraphFailedError")} "${topology.name}"`, errorList);
                         }
                     );
                 }
@@ -82,59 +88,33 @@ export class GraphTopologyItem extends vscode.TreeItem {
         }
     }
 
-    public deleteGraphCommand() {
+    public async deleteGraphCommand() {
         if (this._graphTopology) {
-            // TODO we might need a confirmation before delete
-            GraphTopologyData.deleteGraphTopology(this.iotHubData, this.deviceId, this.moduleId, this._graphTopology.name).then(
-                (response) => {
-                    TreeUtils.refresh();
-                },
-                (error) => {
-                    this._logger.logError(`Failed to delete the graph "${this._graphTopology.name}"`, error); // TODO. localize
-                }
-            );
+            const allowDelete = await ExtensionUtils.showConfirmation(Localizer.localize("deleteGraphConfirmation"));
+            if (allowDelete) {
+                GraphTopologyData.deleteGraphTopology(this.iotHubData, this.deviceId, this.moduleId, this._graphTopology.name).then(
+                    (response) => {
+                        TreeUtils.refresh();
+                    },
+                    (error) => {
+                        const errorList = GraphEditorPanel.parseDirectMethodError(error);
+                        this._logger.logError(`${Localizer.localize("deleteGraphFailedError")} "${this._graphTopology!.name}"`, errorList);
+                    }
+                );
+            }
         }
     }
 
     public createNewGraphInstanceCommand(context: vscode.ExtensionContext) {
-        const createGraphPanel = GraphEditorPanel.createOrShow(context.extensionPath, Localizer.localize("createNewInstancePageTile"));
-        if (createGraphPanel) {
-            createGraphPanel.waitForPostMessage({
-                name: Constants.PostMessageNames.closeWindow,
-                callback: () => {
-                    createGraphPanel.dispose();
-                }
-            });
-
-            createGraphPanel.setupInitialMessage({ pageType: Constants.PageTypes.instancePage, graphData: this._graphTopology });
-
-            createGraphPanel.setupNameCheckMessage((name) => {
-                return (
-                    this._graphInstances == null ||
-                    this._graphInstances.length == 0 ||
-                    this._graphInstances.filter((graph) => {
-                        return graph.name === name;
-                    }).length === 0
-                );
-            });
-
-            createGraphPanel.waitForPostMessage({
-                name: Constants.PostMessageNames.saveInstance,
-                callback: async (instance: MediaGraphInstance) => {
-                    const graphInstance = new InstanceItem(this.iotHubData, this.deviceId, this.moduleId, this._graphTopology, instance);
-                    graphInstance.saveGraph(createGraphPanel, instance);
-                }
-            });
-
-            createGraphPanel.waitForPostMessage({
-                name: Constants.PostMessageNames.saveAndActivate,
-                callback: async (instance: MediaGraphInstance) => {
-                    const graphInstance = new InstanceItem(this.iotHubData, this.deviceId, this.moduleId, this._graphTopology, instance);
-                    graphInstance.saveGraph(createGraphPanel, instance).then(() => {
-                        graphInstance.activateInstanceCommand(instance.name);
-                    });
-                }
-            });
-        }
+        const graphInstance = new InstanceItem(this.iotHubData, this.deviceId, this.moduleId, this._graphTopology!, undefined, (name) => {
+            return (
+                this._graphInstances == null ||
+                this._graphInstances.length == 0 ||
+                this._graphInstances.filter((instance) => {
+                    return instance.name === name;
+                }).length === 0
+            );
+        });
+        graphInstance.setInstanceCommand(context);
     }
 }
